+++ src/ObjectStorage.ts
@@ -1,8 +1,10 @@
-import { Readable, Duplex } from 'stream';
+import { Readable, Duplex, Stream } from 'stream';
 import getStream from 'get-stream';
 import { StorageClient } from './StorageClient';
 
 export type TransformMiddleware = () => Duplex;
+export type ResponseType = 'stream' | 'json' | 'arraybuffer';
+export const DEFAULT_RESPONSE_TYPE: ResponseType = 'json';
 
 export class ObjectStorage {
   private client: StorageClient;
@@ -36,28 +38,23 @@ export class ObjectStorage {
     return this;
   }
 
-  public async getById(objectId: string): Promise<string> {
+  public async getById(objectId: string, responseType: ResponseType = DEFAULT_RESPONSE_TYPE): Promise<any> {
     const { data } = await this.client.readStream(objectId);
     const stream = this.applyMiddlewares(data, this.reverses);
-    return getStream(stream);
+    return ObjectStorage.getDataByResponseType(stream, responseType);
   }
 
   public async getAllByParams(params: object): Promise<string> {
-    const { data } = await this.client.readAllByParamsAsStream(params);
-    const stream = this.applyMiddlewares(data, this.reverses);
-    return getStream(stream);
+    const res = await this.client.readAllByParamsAsStream(params);
+    return res.data;
   }
 
-  public async deleteOne(objectId: string): Promise<string> {
-    const { data } = await this.client.deleteOne(objectId);
-    const stream = this.applyMiddlewares(data, this.reverses);
-    return getStream(stream);
+  public async deleteOne(objectId: string): Promise<any> {
+    return this.client.deleteOne(objectId);
   }
 
-  public async deleteMany(params: object): Promise<string> {
-    const { data } = await this.client.deleteMany(params);
-    const stream = this.applyMiddlewares(data, this.reverses);
-    return getStream(stream);
+  public async deleteMany(params: object): Promise<any> {
+    return this.client.deleteMany(params);
   }
 
   public async postObject(data: object, headers: object): Promise<string> {
@@ -70,5 +67,14 @@ export class ObjectStorage {
     const resultStream = () => this.applyMiddlewares(this.formStream(data), this.forwards);
     const res = await this.client.updateAsStream(objectId, resultStream, headers);
     return res.data;
+  }
+
+  private static getDataByResponseType(data: Stream, responseType: ResponseType) {
+    switch (responseType) {
+      case 'stream': return data;
+      case 'json': return getStream(data);
+      case 'arraybuffer': return getStream.buffer(data);
+      default: throw new Error(`Response type "${responseType}" is not supported`);
+    }
   }
 }