+++ src/StorageClient.ts
@@ -27,8 +27,8 @@ export class StorageClient {
 
   private static httpsAgent = new https.Agent({ keepAlive: true });
 
-  public constructor(config: { uri: string; jwtSecret: string }, axiosInstance?: AxiosInstance) {
-    this.api = axiosInstance || axios.create({
+  public constructor(config: { uri: string; jwtSecret: string }) {
+    this.api = axios.create({
       baseURL: config.uri,
       httpAgent: StorageClient.httpAgent,
       httpsAgent: StorageClient.httpsAgent,
@@ -75,12 +75,11 @@ export class StorageClient {
     return { Authorization: `Bearer ${this.jwtSecret}`, ...override };
   }
 
-  public async readStream(objectId: string, params?: object): Promise<AxiosResponse> {
+  public async readStream(objectId: string): Promise<AxiosResponse> {
     return this.requestRetry(
       async (): Promise<AxiosResponse> => this.api.get(`/objects/${objectId}`, {
         responseType: 'stream',
         headers: await this.getHeaders(),
-        params,
       }),
     );
   }
@@ -88,7 +87,6 @@ export class StorageClient {
   public async readAllByParamsAsStream(params: object): Promise<AxiosResponse> {
     return this.requestRetry(
       async (): Promise<AxiosResponse> => this.api.get('/objects', {
-        responseType: 'stream',
         headers: await this.getHeaders(),
         params,
       }),
@@ -110,7 +108,6 @@ export class StorageClient {
   public async deleteOne(objectId: string): Promise<AxiosResponse> {
     return this.requestRetry(
       async (): Promise<AxiosResponse> => this.api.delete(`/objects/${objectId}`, {
-        responseType: 'stream',
         headers: await this.getHeaders(),
       }),
     );
@@ -119,7 +116,6 @@ export class StorageClient {
   public async deleteMany(params: object): Promise<AxiosResponse> {
     return this.requestRetry(
       async (): Promise<AxiosResponse> => this.api.delete('/objects', {
-        responseType: 'stream',
         headers: await this.getHeaders(),
         params,
       }),