+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 3.4.0 (July 23, 2021)
+# 3.4.1 (July 28, 2021)
+* `ObjectStorage` method `getById` now supports content-types
+* `ObjectStorageWrapper` method `lookupObjectById` now supports content-types
+* Added additional validation for headers
+
+# 3.4.0 (July 22, 2021)
 * Added deleteMany method
 * Added ability to use custom axios config
 * Added ability to add/modify `query` and `meta` headers when updating an object